requirements.txt
pyspark
pyarrow
minio
boto3
python-dotenv
pandas
pytest
pytest-cov
pyyaml